+++ public/js/info-box.js
@@ -17,24 +17,31 @@ function showInfoBox($li, id) {
   $li.appendChild(
     parseHTML(`
     <div class="info-box position-relative overflow=hidden">
-      <div class="route-config">${fieldSet(routeConfig, routeConfig.id)}</div>
+      <div id="loader-${id.replace(/\=/g, "")}" class="backdrop d-none">
+        <div class="spinner-border" role="status">
+          <span class="visually-hidden">Loading...</span>
+        </div>
+      </div>
+      <div class="route-config p-1">${fieldSet(routeConfig, routeConfig.id)}</div>
       <div class="actions justify-content-end p-2" style="display: ${routeConfig._isDefault ? 'none' : 'flex'}">
         <span role="button" class="px-2 pe-1 action-icon" title="reset" onclick="reset('${routeConfig.id}')"><i class="fas fa-undo"></i></span>
         <span role="button" class="px-2 pe-1 action-icon" title="edit" onclick="openModal(this)" data-type="update" data-id="${routeConfig.id}"><i class="fas fa-pen"></i></span>
         <span role="button" class="px-2 pe-1 action-icon" title="clone" onclick="openModal(this)" data-type="clone" data-id="${routeConfig.id}"><i class="fas fa-clone"></i></span>
         <span role="button" class="px-2 pe-1 action-icon" title="refresh" onclick="refresh('${routeConfig.id}')"><i class="fas fa-sync-alt"></i></span>
-        
-        </div>
-        </div>`)
-        );
-      }
-      // <button type="button" class="btn btn-outline-primary box-shadow-none btn-sm ms-2" onclick="reset('${routeConfig.id}')">Reset</button>
-      // <button type="button" class="btn btn-outline-primary box-shadow-none btn-sm ms-2" onclick="openModal(this)" data-type="update" data-id="${routeConfig.id}">Edit</button>
-      // <button type="button" class="btn btn-outline-primary box-shadow-none btn-sm ms-2" onclick="openModal(this)" data-type="clone" data-id="${routeConfig.id}">Clone</button>
-      // <button type="button" class="btn btn-outline-primary box-shadow-none btn-sm ms-2" onclick="refresh('${routeConfig.id}')">Refresh</button>
+      </div>
+    </div>`
+    )
+  );
+}
 
 async function reset(id) {
+  // show info box loader
+  const $infoBoxLoader = document.querySelector(`.info-box #loader-${id.replace(/\=/g, "")}`);
+  $infoBoxLoader.classList.remove("d-none");
+  $infoBoxLoader.classList.add("d-block");
+
   const restoredRoutes = await window.fetch(localhost + "/_reset/db/" + id).then((res) => res.json());
+
   const [routePath, routeConfig] = Object.entries(restoredRoutes)[0];
   resources[routePath] = routeConfig
   toggleInfoBox(id);
@@ -43,7 +50,13 @@ async function reset(id) {
 }
 
 async function refresh(id) {
+  // show info box loader
+  const $infoBoxLoader = document.querySelector(`.info-box #loader-${id.replace(/\=/g, "")}`);
+  $infoBoxLoader.classList.remove("d-none");
+  $infoBoxLoader.classList.add("d-block");
+
   const refreshedRoute = await window.fetch(localhost + "/_db/" + id).then((res) => res.json());
+
   const [routePath, routeConfig] = Object.entries(refreshedRoute)[0];
   resources[routePath] = routeConfig;
   toggleInfoBox(id);
@@ -53,14 +66,25 @@ async function refresh(id) {
 
 function fieldSet(obj, id) {
   return Object.entries(orderRouteConfig(obj)).map(([key, val]) => {
-    if (key === "fetchData" && val) {
-      return `<div class="row px-3"><fieldset><legend>Fetch Data</legend>${fieldSet(val, id)}</fieldset></div>`
+    if (key === "fetchData" && Object.keys(val || {}).length) {
+      return `
+      <div class="row px-3">
+          <label class="key col col-form-label p-0">
+            <button class="btn btn-white fw-semibold border-0 p-0 shadow-none" type="button" 
+            data-bs-toggle="collapse" data-bs-target="#id-${id.replace(/\=/g, "")}_${key}" 
+            aria-expanded="false" aria-controls="id-${id.replace(/\=/g, "")}_${key}">
+              ${key} <i class="fa-solid fa-caret-right"></i>
+            </button>   
+          </label>
+          <div class="val col-12 collapse ps-0" id="id-${id.replace(/\=/g, "")}_${key}">${fieldSet(val, id)}</div>
+      </div>`
     } else {
       return getKeyVal(key, val, id);
     }
   }).join("")
 }
 
+// Todo: Need to optimize the code
 function getKeyVal(key, val, id) {
   if ((!['mock', 'response'].includes(key) && !(val + '')?.length) || val === null || val === undefined || key === "_config") return '';
 
@@ -68,7 +92,7 @@ function getKeyVal(key, val, id) {
     if (!['mock', 'response'].includes(key) && !val.length) return '';
     return `
       <div class="row px-3">
-        <label for="inputEmail3" class="key col col-form-label p-0">${key} :</label>
+        <label class="key col col-form-label p-0">${key} :</label>
         <div class="val col d-flex flex-wrap align-items-center" style="grid-gap:.5rem">
         ${getBadges(val).join("")}
         </div>
@@ -86,20 +110,22 @@ function getKeyVal(key, val, id) {
     }
     return `
     <div class="row px-3">
-      <label for="inputEmail3" class="key col col-form-label p-0 mb-2 w-100" style="max-width: 100%">
-      <button class="btn btn-white text-primary p-0 box-shadow-none" type="button" 
-      data-bs-toggle="collapse" data-bs-target="#id-${id.replace(/\=/g, "")}_${key}" 
-      aria-expanded="false" aria-controls="id-${id.replace(/\=/g, "")}_${key}">
-        ${key} :
-      </button>  
-      <div class="val col-12 collapse" id="id-${id.replace(/\=/g, "")}_${key}">
-        <pre class="form-control">${typeof val === 'object' ? JSON.stringify(val, null, 2) : val}</pre>
+      <label class="key col col-form-label p-0 w-100" style="max-width: 100%">
+        <button class="btn btn-white fw-semibold border-0 p-0 shadow-none" type="button" 
+        data-bs-toggle="collapse" data-bs-target="#id-${id.replace(/\=/g, "")}_${key}" 
+        aria-expanded="false" aria-controls="id-${id.replace(/\=/g, "")}_${key}">
+          ${key} <i class="fa-solid fa-caret-right"></i>
+        </button>  
+      </label>
+      <div class="val col-12 collapse p-0 mt-2 position-relative" id="id-${id.replace(/\=/g, "")}_${key}">
+        <span id="fetch-badge" style="top: -1.8rem; right: 1rem;" class="position-absolute badge bg-secondary ms-4">${typeof val === 'object' ? "JSON" : "STRING"}</span>
+        <textarea class="form-control" rows="5">${typeof val === 'object' ? JSON.stringify(val, null, 2) : val}</textarea>
       </div>
     </div>`;
   } else if ((val + "").indexOf("<img") >= 0) {
     return `
     <div class="row px-3">
-      <label for="inputEmail3" class="key col col-form-label p-0">${key} :</label>
+      <label class="key col col-form-label p-0">${key} :</label>
       <div class="val col">
         <div class="img">${val}</div>
       </div>
@@ -108,7 +134,7 @@ function getKeyVal(key, val, id) {
     if (!(val + "").trim().length) return '';
     return `
       <div class="row px-3">
-        <label for="inputEmail3" class="key col col-form-label p-0">${key} :</label>
+        <label class="key col col-form-label p-0">${key} :</label>
         <div class="val col">${val}</div>
       </div>`;
   }