+++ samples/server.js
@@ -1,7 +1,7 @@
 const { MockServer } = require("@r35007/mock-server");
 
 // Provide config as a param. If not provided, It uses the default Config.
-const mockServer = MockServer.Create();
+const mockServer = MockServer.Create({root:__dirname});
 
 const app = mockServer.app; // Gives you the Express app